确保劳务费汇总里的管理费跟对应开票计划里的一致

+++ HRSystem/src/com/nathan/model/Payroll.java
@@ -189,7 +189,7 @@ public class Payroll {
 	/**
 	 * @param administrationExpenses the administrationExpenses to set
 	 */
-	public void setAdministrationExpenses(float administrationExpenses) {
+	public void setAdministrationExpenses(double administrationExpenses) {
 		this.administrationExpenses = administrationExpenses;
 	}
